+++ frontend/src/components/buttons/DownloadTemplateButton.jsx
@@ -1,0 +1,22 @@
+import Button from "react-bootstrap/Button";
+
+function DownloadTemplateButton({ editor }) {
+  const downloadTemplate = () => {
+    const jsonString = JSON.stringify(editor.getValue(), null, 2);
+    const blob = new Blob([jsonString], { type: "application/json" });
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement("a");
+    link.href = url;
+    link.download = "cyclemetry_template.json";
+    link.click();
+    URL.revokeObjectURL(url);
+  };
+
+  return (
+    <Button variant="primary" className="ms-3" onClick={downloadTemplate}>
+      Download Template
+    </Button>
+  );
+}
+
+export default DownloadTemplateButton;
